# openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/inventory/v1beta1/resources/rhel-hosts:
        put:
            tags:
                - KesselRhelHostService
            operationId: KesselRhelHostService_UpdateRhelHost
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/kessel.inventory.v1beta1.resources.UpdateRhelHostRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/kessel.inventory.v1beta1.resources.UpdateRhelHostResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
        post:
            tags:
                - KesselRhelHostService
            operationId: KesselRhelHostService_CreateRhelHost
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/kessel.inventory.v1beta1.resources.CreateRhelHostRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/kessel.inventory.v1beta1.resources.CreateRhelHostResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
        delete:
            tags:
                - KesselRhelHostService
            operationId: KesselRhelHostService_DeleteRhelHost
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/kessel.inventory.v1beta1.resources.DeleteRhelHostRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/kessel.inventory.v1beta1.resources.DeleteRhelHostResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
        patch:
            tags:
                - KesselRhelHostService
            operationId: KesselRhelHostService_PatchRhelHost
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/kessel.inventory.v1beta1.resources.PatchRhelHostRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/kessel.inventory.v1beta1.resources.PatchRhelHostResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
components:
    schemas:
        google.protobuf.Any:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        google.rpc.Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/google.protobuf.Any'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...
        kessel.inventory.v1beta1.resources.CreateRhelHostRequest:
            type: object
            properties:
                rhel_host:
                    allOf:
                        - $ref: '#/components/schemas/kessel.inventory.v1beta1.resources.RhelHost'
                    description: The Rhel Host to create in Kessel Asset Inventory
        kessel.inventory.v1beta1.resources.CreateRhelHostResponse:
            type: object
            properties: {}
...
        kessel.inventory.v1beta1.resources.DeleteRhelHostRequest:
            type: object
            properties:
                reporter_data:
                    allOf:
                        - $ref: '#/components/schemas/kessel.inventory.v1beta1.resources.ReporterData'
                    description: "The resource to be updated will be defined by \n \\\"<reporter_data.reporter_type>:<reporter_instaance_id>:<reporter_data.local_resource_id>\\\"\n from the request body."
        kessel.inventory.v1beta1.resources.DeleteRhelHostResponse:
            type: object
            properties: {}
...
        kessel.inventory.v1beta1.resources.Metadata:
            type: object
            properties:
                id:
                    readOnly: true
                    type: string
                    description: Kessel Asset Inventory generated identifier.
                resource_type:
                    readOnly: true
                    type: string
                    description: The type of the Resource
                created_at:
                    readOnly: true
                    type: string
                    description: Date and time when the inventory item was first reported.
                    format: date-time
                updated_at:
                    readOnly: true
                    type: string
                    description: Date and time when the inventory item was last updated.
                    format: date-time
                deleted_at:
                    readOnly: true
                    type: string
                    description: Date and time when the inventory item was deleted.
                    format: date-time
                org_id:
                    type: string
                    description: |-
                        The org id in which this resource is a member for access control.  A
                         resource can only be a member of one org.
                workspace_id:
                    type: string
                    description: |-
                        The workspace id in which this resource is a member for access control.  A
                         resource can only be a member of one workspace.
                labels:
                    type: array
                    items:
                        $ref: '#/components/schemas/kessel.inventory.v1beta1.resources.ResourceLabel'
...
        kessel.inventory.v1beta1.resources.PatchRhelHostRequest:
            type: object
            properties:
                rhel_host:
                    $ref: '#/components/schemas/kessel.inventory.v1beta1.resources.RhelHost'
                updateMask:
                    type: string
                    description: |-
                        list of paths that can be patched.
                         Current supported values: rhel_host.metadata.workspace_id, rhel_host.reporter_data.api_href, rhel_host.reporter_data.console_href
                    format: field-mask
        kessel.inventory.v1beta1.resources.PatchRhelHostResponse:
            type: object
            properties: {}
        kessel.inventory.v1beta1.resources.ReporterData:
            type: object
            properties:
                reporter_type:
                    enum:
                        - REPORTER_TYPE_UNSPECIFIED
                        - REPORTER_TYPE_OTHER
                        - ACM
                        - HBI
                        - OCM
                        - NOTIFICATIONS
                    type: string
                    format: enum
                reporter_instance_id:
                    readOnly: true
                    type: string
                    description: |-
                        The ID of the instance of the reporter. This is derived from the
                         authentication mechanism, i.e. authentication token.
                console_href:
                    type: string
                    description: |-
                        The URL for this resource in the reporter's management UI console. For
                         example this would be the cluster URL in the HCC Console for an OCM
                         reported cluster.
                api_href:
                    type: string
                    description: Reporter specific API link to the resource.
                local_resource_id:
                    type: string
                    description: |-
                        The ID assigned to this resource by the reporter, for example OCM cluster
                         ID, HBI's host id, or ACM managed cluster name etc.
                reporter_version:
                    type: string
                    description: version of the reporter
        kessel.inventory.v1beta1.resources.ResourceLabel:
            type: object
            properties:
                key:
                    type: string
                value:
                    type: string
        kessel.inventory.v1beta1.resources.RhelHost:
            type: object
            properties:
                metadata:
                    allOf:
                        - $ref: '#/components/schemas/kessel.inventory.v1beta1.resources.Metadata'
                    description: Metadata about this resource
                reporter_data:
                    writeOnly: true
                    allOf:
                        - $ref: '#/components/schemas/kessel.inventory.v1beta1.resources.ReporterData'
                    description: Write only reporter specific data
...
        kessel.inventory.v1beta1.resources.UpdateRhelHostRequest:
            type: object
            properties:
                rhel_host:
                    allOf:
                        - $ref: '#/components/schemas/kessel.inventory.v1beta1.resources.RhelHost'
                    description: "The resource to be updated will be defined by \n \\\"<reporter_data.reporter_type>:<reporter_instaance_id>:<reporter_data.local_resource_id>\\\"\n from the request body."
        kessel.inventory.v1beta1.resources.UpdateRhelHostResponse:
            type: object
            properties: {}
...